+++ src/combinator/control/constraint/contract.ts
@@ -1,4 +1,4 @@
-import { Parser, Data, eval, exec, check } from '../../data/parser';
+import { Parser, Data, Config, eval, exec, check } from '../../data/parser';
 
 export function contract<P extends Parser<unknown>>(patterns: string | RegExp | (string | RegExp)[], parser: P, cond: (results: readonly Data<P>[], rest: string) => boolean): P;
 export function contract<T, D extends Parser<unknown>[]>(patterns: string | RegExp | (string | RegExp)[], parser: Parser<T, D>, cond: (results: readonly T[], rest: string) => boolean): Parser<T, D> {
@@ -31,15 +31,15 @@ export function validate<T, D extends Parser<unknown>[]>(patterns: string | RegE
   }
 }
 
-export function verify<P extends Parser<unknown>>(parser: P, cond: (results: readonly Data<P>[], rest: string) => boolean): P;
-export function verify<T, D extends Parser<unknown>[]>(parser: Parser<T, D>, cond: (results: readonly T[], rest: string) => boolean): Parser<T, D> {
+export function verify<P extends Parser<unknown>>(parser: P, cond: (results: readonly Data<P>[], rest: string, config: Config<P>) => boolean): P;
+export function verify<T, D extends Parser<unknown, any, S, C>[], S extends object, C extends object>(parser: Parser<T, D, S, C>, cond: (results: readonly T[], rest: string, config: C) => boolean): Parser<T, D, S, C> {
   assert(parser);
   return (source, config) => {
     if (source === '') return;
     const result = parser(source, config);
     assert(check(source, result));
     if (!result) return;
-    if (!cond(eval(result), exec(result))) return;
+    if (!cond(eval(result), exec(result), config)) return;
     return exec(result).length < source.length
       ? result
       : undefined;