+++ src/parser/autolink.ts
@@ -1,13 +1,13 @@
 import { MarkdownParser } from '../../markdown';
-import { union } from '../combinator';
+import { union, lazy } from '../combinator';
 import { autolink as autolink_ } from './inline';
 import { newline, unescsource } from './source';
 
 export import Config = MarkdownParser.Config;
 export import AutolinkParser = MarkdownParser.AutolinkParser;
 
-export const autolink: AutolinkParser = union([
+export const autolink: AutolinkParser = lazy(() => union([
   autolink_,
   newline,
   unescsource
-]);
+]));